+++ docs/ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## PiPedal 1.3.60 Release
+
+Bug fixes:
+- Uploading of preset and bank files now  work properly.
+- Uploaded plugin presets are merged with existing presets, instead of overwriting existing presets.
+
 ## PiPedal 1.3.59 Release
 
 20% performance improvement when using Toob Neural Amp Modeller with large Wavenet models.